refactor(services): fix typo in UrlManagementService file name
